requirements.txt
python-telegram-bot==20.7
mistralai==0.0.7
langchain==0.1.0
chromadb==0.4.22
python-dotenv==1.0.0
python-docx-template==0.5.2
huggingface-hub==0.20.3
flask==2.3.2
gunicorn==20.1.0
